Refactor API to be consistent with UI
