+++ expense-authorization/src/API/EMBC.ExpenseAuthorization.Api/Models/EocApproval.cs
@@ -1,0 +1,11 @@
+﻿using System;
+
+namespace EMBC.ExpenseAuthorization.Api.Models
+{
+    public class EocApproval
+    {
+        public string ApprovedBy { get; set; }
+        public string Position { get; set; }
+        public DateTime ApprovalDateTime { get; set; }
+    }
+}
