+++ expense-authorization/src/API/EMBC.ExpenseAuthorization.Api/Models/EocApprovals.cs
@@ -1,0 +1,8 @@
+﻿namespace EMBC.ExpenseAuthorization.Api.Models
+{
+    public class EocApprovals
+    {
+        public EocApproval Processing { get; set; }
+        public EocApproval ExpenditureRequest { get; set; }
+    }
+}
