+++ expense-authorization/src/API/EMBC.Tests.Unit.ExpenseAuthorization.Api/Controllers/ResourceRequestControllerTests.cs
@@ -1,8 +1,8 @@
 ﻿using System.Threading;
 using System.Threading.Tasks;
 using EMBC.ExpenseAuthorization.Api.Controllers;
-using EMBC.ExpenseAuthorization.Api.ETeam.Models;
 using EMBC.ExpenseAuthorization.Api.Features;
+using EMBC.ExpenseAuthorization.Api.Models;
 using MediatR;
 using Microsoft.AspNetCore.Mvc;
 using Moq;
@@ -25,9 +25,9 @@ namespace EMBC.Tests.Unit.ExpenseAuthorization.Api.Controllers
                 .Setup(m => m.Send(It.IsAny<ResourceRequest.CreateCommand>(), It.IsAny<CancellationToken>()))
                 .ReturnsAsync(response);
 
-            var sut = new ResourceRequestController(mockMediator.Object, mockLogger.Object);
+            var sut = new ExpenseAuthorizationController(mockMediator.Object, mockLogger.Object);
 
-            var actual = await sut.PostAsync(new ResourceRequestModel(), null);
+            var actual = await sut.PostAsync(new ExpenseAuthorizationRequest(), null);
 
             Assert.IsType<OkResult>(actual);
